chore(tests): silence console.{log|warn|error} messages during tests
this adds a little utility for replacing the `console` functions with
mocks to silence any error messages and get our jest output under
control.

This function calls `afterAll` internally, so it will clean up after
itself by restoring the original functions for `console.log`, `.error`,
etc.
