+++ src/runtime/test/hydrate-prop-types.spec.tsx
@@ -1,7 +1,14 @@
 import { Component, Host, Prop, h } from '@stencil/core';
 import { newSpecPage } from '@stencil/core/testing';
+import { setupConsoleMocker } from '../../testing/testing-utils';
 
 describe('hydrate prop types', () => {
+  const consoleMockSetup = setupConsoleMocker();
+
+  beforeEach(() => {
+    consoleMockSetup();
+  });
+
   it('number', async () => {
     @Component({ tag: 'cmp-a' })
     class CmpA {
